function for changing color of bar price range
