+++ components/price-range/ExpenseChart.tsx
@@ -1,131 +1,45 @@
 import React from "react";
 
-interface barProps {
+interface BarProps {
+  totalBars: number;
   disabledBefor: number;
   disabledAfter: number;
 }
+type Datas = number[];
+let datas: Datas;
 
-const BarChart: React.FC<barProps> = ({ disabledBefor, disabledAfter }) => {
-  let datas = [
-    "22%",
-    "50%",
-    "20%",
-    "30%",
-    "40%",
-    "40%",
-    "4%",
-    "3%",
-    "4%",
-    "5%",
-    "22%",
-    "50%",
-    "20%",
-    "30%",
-    "40%",
-    "40%",
-    "4%",
-    "3%",
-    "4%",
-    "5%",
-    "22%",
-    "50%",
-    "20%",
-    "30%",
-    "40%",
-    "40%",
-    "4%",
-    "3%",
-    "4%",
-    "5%",
-    "22%",
-    "50%",
-    "20%",
-    "30%",
-    "40%",
-    "40%",
-    "4%",
-    "3%",
-    "4%",
-    "5%",
-    "22%",
-    "50%",
-    "20%",
-    "30%",
-    "40%",
-    "40%",
-    "4%",
-    "3%",
-    "4%",
-    "5%",
-    "22%",
-    "50%",
-    "20%",
-    "30%",
-    "40%",
-    "40%",
-    "4%",
-    "3%",
-    "4%",
-    "5%",
-    "22%",
-    "50%",
-    "20%",
-    "30%",
-    "40%",
-    "40%",
-    "4%",
-    "3%",
-    "4%",
-    "5%",
-    "22%",
-    "50%",
-    "20%",
-    "30%",
-    "40%",
-    "40%",
-    "4%",
-    "3%",
-    "4%",
-    "5%",
-    "22%",
-    "50%",
-    "20%",
-    "30%",
-    "40%",
-    "40%",
-    "4%",
-    "3%",
-    "4%",
-    "5%",
-    "22%",
-    "50%",
-    "20%",
-    "30%",
-    "40%",
-    "40%",
-    "4%",
-    "3%",
-    "4%",
-    "5%",
+const BarChart: React.FC<BarProps> = ({
+  totalBars,
+  disabledBefor,
+  disabledAfter,
+}) => {
+  datas = [
+    2, 5, 4, 7, 12, 16, 17, 3, 32, 22, 22, 55, 26, 28, 35, 40, 44, 48, 52, 70,
+    72, 74, 76, 78, 80, 40, 38, 36, 34, 34, 22, 50, 20, 30, 40, 40, 4, 28, 26,
+    24, 22, 18, 16, 14, 12, 10, 8, 6, 4, 2,
   ];
+
   return (
     <div className="w-full h-36 flex justify-around items-end">
-      {datas.map((data, index) =>
-        index < disabledBefor || index >= disabledAfter ? (
-          <span
-            key={index}
-            className={`w-2 bg-gray-200 rounded-t-sm`}
-            style={{ height: data }}
-          ></span>
-        ) : (
-          <span
-            key={index}
-            className={`w-2 bg-[#E31C5F] rounded-t-sm`}
-            style={{ height: data }}
-          ></span>
-        )
+      {datas.length == totalBars ? (
+        datas.map((data, index) => {
+          const isDisabled = index < disabledBefor || index >= disabledAfter;
+          // console.log(isDisabled);
+          return (
+            <div
+              key={index}
+              className={`w-2 rounded-t-sm h-full ${
+                isDisabled ? "bg-[#8a8a8a]" : "bg-[#E31C5F]"
+              }`}
+              style={{ height: `${data}%` }}
+            ></div>
+          );
+        })
+      ) : (
+        <div></div>
       )}
     </div>
   );
 };
+
 export default BarChart;
